requirements.txt
pymorphy2~=0.9.1
pymongo==4.4.1
pandas==2.0.3
SQLAlchemy==2.0.19
psutil==5.9.5
scikit-learn==1.3.0
numpy==1.25.1
torch==2.0.1
matplotlib==3.7.2
plotly==5.15.0
pytz~=2023.3
python-dotenv~=1.0.0
Pyrogram==2.0.106
scipy==1.11.1
tinkoff~=0.1.1